refactor(ui): InputWithValidation 디렉토리 구조로 변경
InputWithValidation 컴포넌트를 디렉토리 기반
구조로 리팩토링했습니다.

- 파일 형태에서 디렉토리 구조로 변경
- import 경로 수정 (상대 경로 조정)
- Input 컴포넌트 색상을 하드코딩에서 theme vars로 변경
