+++ packages/ui/src/components/molecules/inputWithValidation/inputWithValidation.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import * as SvgIcons from '../../assets/svgs';
+import * as SvgIcons from '../../../assets/svgs';
 import InputWithValidation from './InputWithValidation';
 
 const meta = {
